chore(dailyFragments): move logic to service
