+++ src/modules/expeditions/models/Visit.model.ts
@@ -47,3 +47,4 @@ VisitSchema.plugin(MongooseDelete, {
 
 // register the model and export it
 export const VisitModel = model<IVisit>('Visit', VisitSchema);
+export type VisitModel = typeof VisitModel;
